requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
Django==3.2.6
django-bootstrap4==3.0.1
django-braces==1.14.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
importlib-metadata==2.1.1
Pillow==8.3.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
wincertstore==0.2
zipp==3.5.0
